pyproject.toml
[tool.poetry]
name = "epivar"
version = "0.1.0"
description = ""
authors = ["Jan Binkowski <jan.binkowski@pum.edu.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.2.3"
fastapi = "^0.115.13"
django-filter = "^25.1"
django-encrypted-model-fields = "^0.6.5"
psycopg2 = "^2.9.10"
dotenv = "^0.9.9"
click = "^8.2.1"
requests = "^2.32.4"
loguru = "^0.7.3"
sqlalchemy = "^2.0.41"
tqdm = "^4.67.1"
django-formtools = "^2.5.1"
pronto = "^2.7.0"
pandas = "^2.3.0"
celery = {extras = ["redis"], version = "^5.5.3"}
pybedtools = "^0.12.0"
django-celery-results = "^2.6.0"
django-reversion = "^5.1.0"
gseapy = "^1.1.9"
pgvector = "^0.4.1"
sentence-transformers = "^5.0.0"
pytabix = "^0.1"
django-polymorphic = "^4.1.0"
pyarrow = "^21.0.0"
pydantic = "^2.11.7"
statsmodels = "^0.14.5"
plotly = "^6.3.0"
dash-bio = "^1.0.2"
liftover = "^1.3.2"
django-crispy-forms = "^2.4"
crispy-bootstrap5 = "^2025.6"
docutils = "^0.22"
django-select2 = "^8.4.1"
pyranges = "^0.1.4"
django-addanother = "^2.2.2"
gunicorn = "^23.0.0"
django-tinymce = "^4.1.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
